services/ms-notification/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
aio-pika>=9.4,<10.0
prometheus-client>=0.20,<1.0
aiosmtplib>=3.0,<4.0
cassandra-driver>=3.29,<4.0
